% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accounts.R
\name{ocbc_accounts}
\alias{ocbc_accounts}
\title{Make requests to OCBC Accounts category APIs}
\usage{
ocbc_accounts(
  account_type = c("Children", "Deposit", "Foreign", "Saving", "Current", "Trading"),
  acctoken,
  country = NULL,
  subCategory = NULL,
  productName = NULL,
  benefit = NULL
)
}
\arguments{
\item{account_type}{The account type, hence the API that the function will call from. Required. Options are Children, Deposit, Foreign, Saving, Current, Trading. Note: it is recommended to only specify 1 account type at a time.}

\item{acctoken}{Account token. Required. Users can sign up for their api key, api secret and account token at https://api.ocbc.com/store/home}

\item{country}{Either "SG" for Singapore, or "MY" for Malaysia. Optional. Note: Deposit, Foreign and Trading Accounts do not have the country parameter. If specified, the country parameter is automatically ignored.}

\item{subCategory}{Filter accounts by subcategories. Optional. (e.g Children's account, Foreign Currency) Note: subCategory should correspond to the account_type (e.g Children's account for account_type = 'Children')}

\item{productName}{Filter accounts by keywords. Optional. Note: productName should correspond to the account_type (e.g Mighty Savers, Child Development for account_type = 'Children')}

\item{benefit}{Filter accounts by benefit. Optional. (e.g Bonus Interest, Interest)}
}
\value{
A detailed dataframe of OCBC accounts with variables such as eligibility, minimumBalance, feesAndcharges, benefits, providing the user with the full information about the accounts.
}
\description{
This function makes the information from the OCBC Accounts category APIs easily accessible in an R-readable format.
There are 6 different types of accounts, each with their own unique API \cr
1. Children's Accounts \cr
2. Deposit Accounts \cr
3. Foreign Accounts \cr
4. Saving Accounts \cr
5. Current Accounts \cr
6. Trading Accounts \cr
}
\examples{
ocbc_accounts("Children", acctoken = Sys.getenv("ACC_TOKEN_OCBC"),
country = "SG", productName = "Mighty Savers")
}
