% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cards.R
\name{ocbc_cards}
\alias{ocbc_cards}
\title{Make requests to OCBC Cards category APIs}
\usage{
ocbc_cards(
  card_type = c("Credit", "Credit Advisor", "Credit Promotions ", "Debit Advisor",
    "Rewards"),
  acctoken,
  country = NULL,
  product_type = NULL,
  keyword = NULL,
  tag = NULL,
  name = NULL,
  address = NULL,
  page = NULL,
  limit = 10,
  category = NULL,
  merchant = NULL,
  price = NULL
)
}
\arguments{
\item{card_type}{The card category, hence the API that the function will call from. Required.  Options are Credit, Credit Advisor, Credit Promotions, Debit Advisor and Rewards. Note: it is recommended to only specify 1 card type at a time.}

\item{acctoken}{Account token. Required. Users can sign up for their api key, api secret and account token at https://api.ocbc.com/store/home}

\item{country}{Country for which available credit cards products are retrieved. Currently supported only for Malaysia (MY). Optional. Note: Parameter for 'Credit' category only. If specified for other categories, this parameter will be automatically ignored.}

\item{product_type}{Product type. Optional. Note: Parameter for 'Credit' category only. If specified for other categories, this parameter will be automatically ignored.}

\item{keyword}{Recommend a credit or debit card based on this keyword. Optional. Note: Parameter for 'Credit Advisor' and 'Debit Advisor categories only.}

\item{tag}{Filter promotions by categories and subcategories. Optional. Note: Parameter for 'Credit Promotions category only.}

\item{name}{Filter promotions by merchant names. Optional. Note: Parameter for 'Credit Promotions category only.}

\item{address}{Filter promotions by address. Optional. Note: Parameter for 'Credit Promotions category only.}

\item{page}{Specify page number for paginated response. Optional. Note: Parameter for 'Credit Promotions' and 'Rewards Catalogue' category only.}

\item{limit}{Number of promotion entries to return per page. Default = 10. Optional. Note: Parameter for 'Credit Promotions' and 'Rewards Catalogue' category only.}

\item{category}{Filter rewards by categories (e.g: retail, dining, entertainment, travel, other). Optional. Note: Parameter for 'Rewards Catalogue' category only.}

\item{merchant}{Filter rewards by merchant names. Optional. Note: Parameter for 'Rewards Catalogue' category only.}

\item{price}{Filter maximum OCBC$ cost of rewards. Optional. Note: Parameter for 'Rewards Catalogue' category only.}
}
\value{
A detailed dataframe of OCBC card categories with information related to the category specified (e.g A dataframe with information such as merchant, redemption item and validity date of each reward). For the 'Credit Promotions' category, the function returns a list of 2 dataframes: one of the information of the promotions and the other the terms and conditions.
}
\description{
This function makes the information from the OCBC Cards category APIs easily accessible in an R-readable format.
There are 5 different categories, each with their own unique API \cr
1. Credit Cards \cr
2. Credit Card Advisor \cr
3. Credit Card Promotions \cr
4. Debit Card Advisor \cr
5. Rewards Catalogue \cr
}
\examples{
ocbc_cards("Rewards", acctoken = Sys.getenv("ACC_TOKEN_OCBC"), limit = "30")
}
