% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Insurance.R
\name{ocbc_insurance}
\alias{ocbc_insurance}
\title{Make requests to OCBC Insurance category APIs}
\usage{
ocbc_insurance(
  insurance_type = c("Accident and Health", "Car", "Endowment", "Home and Mortgage",
    "Life", "Maternity", "Travel"),
  acctoken,
  subCategory = NULL,
  productName = NULL,
  benefit = NULL
)
}
\arguments{
\item{insurance_type}{The insurance type, hence the API that the function will call from. Required. Options are Accident and Health, Car, Endowment, Home and Mortgage, Life, Maternity and Travel. Note: it is recommended to only specify 1 insurance type at a time.}

\item{acctoken}{Account token. Required. Users can sign up for their api key, api secret and account token at https://api.ocbc.com/store/home}

\item{subCategory}{Filter policies by subcategory. Optional.}

\item{productName}{Filter policies by product name. Optional.}

\item{benefit}{Filter policies by benefit. Optional.}
}
\value{
A detailed dataframe of OCBC insurance policies with variables such as product name, benefits, premiums, eligibility etc.
}
\description{
This function makes the information from the OCBC Insurance category APIs easily accessible in an R-readable format.
There are 7 different types of insurance, each with their own unique API \cr
1. Accident and Health Insurance \cr
2. Car Insurance \cr
3. Endowment Insurance \cr
4. Home and Mortgage Insurance \cr
5. Life Insurance \cr
6. Maternity Insurance \cr
7. Travel Insurance \cr
}
\examples{
ocbc_insurance("Accident and Health", acctoken = Sys.getenv("ACC_TOKEN_OCBC"),
productName = "Supreme Health")
}
