% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Investments.R
\name{ocbc_investments}
\alias{ocbc_investments}
\title{Make requests to OCBC Investments category APIs}
\usage{
ocbc_investments(
  investment_type = c("Bonds", "Equity and Forex", "Listed Securities ",
    "Leveraged Forex", "Treasury", "Unlisted Securities", "Unit Trusts"),
  acctoken,
  productName = NULL,
  benefit = NULL
)
}
\arguments{
\item{investment_type}{The investment type, hence the API that the function will call from. Required. Options are Bonds, Equity and Forex, Listed Securities, Leveraged Forex, Treasury, Unlisted Securities and Unit Trusts}

\item{acctoken}{Account token. Required. Users can sign up for their api key, api secret and account token at https://api.ocbc.com/store/home}

\item{productName}{Filter investment products by product name. Optional. E.g Bonds}

\item{benefit}{Filter investment products by benefit. Optional. E.g interest}
}
\value{
A detailed dataframe of OCBC investment products with variables such as product name, benefits, risks etc.
}
\description{
This function makes the information from the OCBC Investments category APIs easily accessible in an R-readable format.
There are 7 different types of investments, each with their own unique API \cr
1. Bonds \cr
2. Equity and Forex \cr
3. Listed Securities \cr
4. Leveraged Forex and Futures \cr
5. Treasury \cr
6. Unlisted Securities \cr
7. Unit Trusts
}
\examples{
ocbc_investments("Bonds", acctoken = Sys.getenv("ACC_TOKEN_OCBC"))
}
