% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loans.R
\name{ocbc_loans}
\alias{ocbc_loans}
\title{Make requests to OCBC Loans category APIs}
\usage{
ocbc_loans(
  loan_type = c("Car", "Cash", "Home", "Property", "Secured Overdraft", "SME", "Study"),
  acctoken,
  subCategory = NULL,
  productName = NULL,
  benefit = NULL
)
}
\arguments{
\item{loan_type}{The loan type, hence the API that the function will call from. Required. Options are Car, Cash, Home, Property, Secured Overdraft, SME and Study.}

\item{acctoken}{Account token. Required. Users can sign up for their api key, api secret and account token at https://api.ocbc.com/store/home}

\item{subCategory}{Filter loans by subcategory names. Optional.}

\item{productName}{Filter loans by product name.}

\item{benefit}{Filter loans by benefit.}
}
\value{
A detailed dataframe of OCBC loans with variables such as product name, description, eligbility, benefits, loan amount etc.
}
\description{
This function makes the information from the OCBC Loans category APIs easily accessible in an R-readable format.
There are 8 different types of Loans, each with their own unique API. The 'Loans Accounts List' category requires an OAuth 2.0 authentication, hence only 7 APIs are made available. \cr
1. Car Loans \cr
2. Cash Loans \cr
3. Home Loans \cr
4. Property Loans \cr
5. Secured Overdraft \cr
6. SME Loans \cr
7. Study Loans \cr
}
\examples{
ocbc_loans("Car", acctoken = Sys.getenv("ACC_TOKEN_OCBC"), productName = "New Car Financing")
}
